Debería estar terminado todo lo de jwt

+++ src/routes/PropiedadRoutes.ts
@@ -1,9 +1,12 @@
+/* eslint-disable @typescript-eslint/no-unsafe-member-access,max-len */
 import HttpStatusCodes from '@src/common/HttpStatusCodes';
 
 import PropiedadService from '@src/services/PropiedadService';
 import { IPropiedad } from '@src/models/Propiedad';
 import { IReq, IRes } from './types/express/misc';
 import {IFiltrosPropiedad} from '@src/models/FiltrosPropiedad';
+import Usuario from '@src/models/Usuario';
+import UsuarioService from '@src/services/UsuarioService';
 
 
 // **** Functions **** //
@@ -20,7 +23,15 @@ async function getAll(_: IReq, res: IRes) {
  * get de todas las propiedades de x usuario.
  */
 async function getUsuario(req: IReq, res: IRes) {
-  const idUsuario =+req.params.id;
+  const idUsuario: number =+req.params.id;
+
+  const token: string = (req.headers['authorization'] as string).split(' ')[1];
+  const id_token: number = JSON.parse(atob(token.split('.')[1])).data as number;
+
+  if (!Usuario.isAdmin(await UsuarioService.getOne(id_token)) && idUsuario != id_token) {
+    return res.status(HttpStatusCodes.UNAUTHORIZED);
+  }
+
   const propiedades = await PropiedadService.getUsuario(idUsuario);
   return res.status(HttpStatusCodes.OK).json({ propiedades });
 }
@@ -67,6 +78,14 @@ async function getOne(req: IReq, res: IRes) {
  */
 async function add(req: IReq<{propiedad: IPropiedad}>, res: IRes) {
   const { propiedad } = req.body;
+
+  const token: string = (req.headers['authorization'] as string).split(' ')[1];
+  const id_token: number = JSON.parse(atob(token.split('.')[1])).data as number;
+
+  if (!Usuario.isAdmin(await UsuarioService.getOne(id_token)) && propiedad.duenio != id_token) {
+    return res.status(HttpStatusCodes.UNAUTHORIZED);
+  }
+
   await PropiedadService.addOne(propiedad);
   return res.status(HttpStatusCodes.CREATED).end();
 }
@@ -76,6 +95,14 @@ async function add(req: IReq<{propiedad: IPropiedad}>, res: IRes) {
  */
 async function update(req: IReq<{propiedad: IPropiedad}>, res: IRes) {
   const { propiedad } = req.body;
+
+  const token: string = (req.headers['authorization'] as string).split(' ')[1];
+  const id_token: number = JSON.parse(atob(token.split('.')[1])).data as number;
+
+  if (!Usuario.isAdmin(await UsuarioService.getOne(id_token)) && propiedad.duenio != id_token) {
+    return res.status(HttpStatusCodes.UNAUTHORIZED);
+  }
+
   await PropiedadService.updateOne(propiedad);
   return res.status(HttpStatusCodes.OK).end();
 }
@@ -85,6 +112,14 @@ async function update(req: IReq<{propiedad: IPropiedad}>, res: IRes) {
  */
 async function delete_(req: IReq, res: IRes) {
   const id = +req.params.id;
+
+  const token: string = (req.headers['authorization'] as string).split(' ')[1];
+  const id_token: number = JSON.parse(atob(token.split('.')[1])).data as number;
+
+  if (!Usuario.isAdmin(await UsuarioService.getOne(id_token)) && (await PropiedadService.getOne(id)).duenio != id_token) {
+    return res.status(HttpStatusCodes.UNAUTHORIZED);
+  }
+
   await PropiedadService.delete(id);
   return res.status(HttpStatusCodes.OK).end();
 }
